version 0.8.3 ***Lesson8*** +refactoring ->В процессе создание БД

+++ app/src/main/java/com/example/kinofanplus/model/DetailsRepository.kt
@@ -4,5 +4,5 @@ import com.example.kinofanplus.model.movie_list_gson.MovieDTO
 import retrofit2.Callback
 
 interface DetailsRepository {
-    fun getMovieDetailFromServer(requestID: Int, callback: Callback<MovieDTO>)
+    fun getMovieDetailFromServer(requestID: Long, callback: Callback<MovieDTO>)
 }
