+++ app/src/main/java/com/example/kinofanplus/model/DetailsRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.example.kinofanplus.model.movie_list_gson.MovieDTO
 
 class DetailsRepositoryImpl(private val remoteDataSource: RemoteDataSource) : DetailsRepository {
 
-    override fun getMovieDetailFromServer(requestID: Int, callback: Callback<MovieDTO>) {
+    override fun getMovieDetailFromServer(requestID: Long, callback: Callback<MovieDTO>) {
         remoteDataSource.getMovieDetail(requestID, callback)
     }
 }
